Move all interface related modules to primitives.base sub-package
