+++ qiskit/primitives/__init__.py
@@ -48,17 +48,15 @@ Results
 .. autosummary::
    :toctree: ../stubs/
 
-   BasePrimitiveResult
    EstimatorResult
    SamplerResult
 """
 
-from .base_estimator import BaseEstimator
-from .base_result import BasePrimitiveResult
-from .base_sampler import BaseSampler
+from .base import BaseEstimator
+from .base import BaseSampler
 from .backend_estimator import BackendEstimator
 from .estimator import Estimator
-from .estimator_result import EstimatorResult
+from .base.estimator_result import EstimatorResult
 from .backend_sampler import BackendSampler
 from .sampler import Sampler
-from .sampler_result import SamplerResult
+from .base.sampler_result import SamplerResult
