+++ qiskit/primitives/base/base_result.py
