+++ qiskit/primitives/base/sampler_result.py
