+++ qiskit/primitives/estimator.py
@@ -26,8 +26,8 @@ from qiskit.opflow import PauliSumOp
 from qiskit.quantum_info import Statevector
 from qiskit.quantum_info.operators.base_operator import BaseOperator
 
-from .base_estimator import BaseEstimator
-from .estimator_result import EstimatorResult
+from .base import BaseEstimator
+from .base.estimator_result import EstimatorResult
 from .primitive_job import PrimitiveJob
 from .utils import (
     _circuit_key,
